requirements.txt
requests
beautifulsoup4
pycryptodome
playwright
rich
pyyaml
pytest
matplotlib
# 可选依赖
jsbeautifier